requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.5
django-bulk-update==2.2.0
django-tables2==2.3.4
django-widget-tweaks==1.4.8
idna==2.10
iniconfig==1.1.1
japanize-matplotlib==1.1.3
matplotlib==3.3.4
mysqlclient==2.0.3
numpy==1.19.5
packaging==20.8
pandas==1.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
seaborn==0.11.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
xmltodict==0.12.0
stripe==2.55.1
SQLAlchemy==1.3.23
